19_GitInfo/requirements.txt
flask
PyGithub